Fix test compilation by using @testable import for internal types
Remove TestKit mocks that extended internal types (AddressRecord,
WalletKeyStore, PerpetualsSceneViewModel, PerpetualPortfolioSceneViewModel)
and move mocks into test files with @testable import instead.
Add @testable import to test files accessing internal APIs across
Formatters, Store, Keystore, Signer, FiatConnect, and BannerService.
