+++ Features/Perpetuals/Tests/PerpetualsSceneViewModelTests.swift
@@ -3,6 +3,10 @@
 import Testing
 import Primitives
 import PrimitivesTestKit
+import PerpetualService
+import PerpetualServiceTestKit
+import ActivityService
+import ActivityServiceTestKit
 import PerpetualsTestKit
 @testable import Perpetuals
 
@@ -15,5 +19,20 @@ struct PerpetualsSceneViewModelTests {
         let model = PerpetualsSceneViewModel.mock(wallet: wallet)
 
         #expect(model.headerViewModel.walletType == .multicoin)
+    }
+}
+
+extension PerpetualsSceneViewModel {
+    @MainActor
+    static func mock(
+        wallet: Wallet = .mock(),
+        perpetualService: PerpetualServiceable = PerpetualService.mock(),
+        activityService: ActivityService = .mock()
+    ) -> PerpetualsSceneViewModel {
+        PerpetualsSceneViewModel(
+            wallet: wallet,
+            perpetualService: perpetualService,
+            activityService: activityService
+        )
     }
 }
